% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_treatment.R
\name{time_to_treatment}
\alias{time_to_treatment}
\title{Time to treatment}
\usage{
time_to_treatment(patient_df, treatment_df)
}
\arguments{
\item{patient_df}{data.frame with patient info, must have columns patient_id and diagnosis_date}

\item{treatment_df}{data.frame with treatment info, must have columns patient_id and treatment_date}
}
\value{
patient df added with a time_for_treatment column, arranged by patient_id and time_for_treatment
}
\description{
This function adds a column containing time to treatment in days (starting from diagnosis date)
}
\examples{

data(patients)
data(treatment)
time_to_treatment(patients,treatment)
}
